+++ src/ezq/__about__.py
@@ -3,9 +3,9 @@
 """Simple wrapper for python multiprocessing."""
 
 __url__ = "https://github.com/metaist/ezq"
-__version__ = "2.0.2"
-__pubdate__ = "2021-08-24T17:25:52Z"  # date/time in UTC
+__version__ = "2.0.3"
+__pubdate__ = "2023-05-10T03:44:15Z"  # date/time in UTC
 __author__ = "Metaist LLC"
 __email__ = "metaist@metaist.com"
-__copyright__ = "Copyright 2021 by Metaist LLC"
+__copyright__ = "Copyright 2023 by Metaist LLC"
 __license__ = "MIT"