+++ CHANGELOG.md
@@ -19,6 +19,8 @@ These are changes that are on `main` that are not yet in `production`.
 
 - `sortiter` now sorts the list of waiting messages in place to improve performance (~50%).
 
+---
+
 ## [2.0.2] - 2021-08-24T17:25:52Z
 
 [2.0.2]: https://github.com/metaist/ezq/compare/2.0.1...2.0.2